refactor(shops): split god-file shops.service.ts into domain sub-services
Extract CRUD/settings, appearance upload, members, and audit log logic into
separate @Injectable sub-services under shops/services/; orchestrator
delegates only, no logic — 99 lines down from 860. Public API unchanged.
