+++ backend/src/shops/services/shops-members.service.ts
@@ -1,0 +1,205 @@
+import { Injectable } from '@nestjs/common';
+import { PrismaService } from '../../common/prisma/prisma.service';
+import { ErrorCode, AppException } from '../../common/exceptions/http-exception.filter';
+import { AddShopAdminDto } from '../dto/add-shop-admin.dto';
+import { QueryShopMembersDto } from '../dto/query-shop-members.dto';
+import { ShopAuditAction, ShopRole } from '../../generated/prisma/client';
+import { RolesGuard } from '../../common/guards/roles.guard';
+import { ShopsAuditService } from './shops-audit.service';
+import { ShopsCrudService } from './shops-crud.service';
+import { totalPagesFromCount } from '../../common/utils/pagination.utils';
+import { isUUID } from 'class-validator';
+import * as bcrypt from 'bcrypt';
+
+@Injectable()
+export class ShopsMembersService {
+  constructor(
+    private prisma: PrismaService,
+    private readonly audit: ShopsAuditService,
+    private readonly crud: ShopsCrudService,
+  ) {}
+
+  async findMembers(shopId: string, query: QueryShopMembersDto) {
+    await this.crud.findOne(shopId); // 404 guard
+    const page = query.page || 1;
+    const pageSize = query.page_size || 20;
+    const skip = (page - 1) * pageSize;
+
+    const where = { shop_id: shopId };
+    const [members, total] = await this.prisma.$transaction([
+      this.prisma.userShopRole.findMany({
+        where,
+        include: { user: { select: { id: true, email: true, full_name: true, role: true, is_active: true } } },
+        orderBy: { user_id: 'asc' },
+        skip,
+        take: pageSize,
+      }),
+      this.prisma.userShopRole.count({ where }),
+    ]);
+
+    return {
+      members,
+      pagination: {
+        page,
+        page_size: pageSize,
+        total,
+        total_pages: totalPagesFromCount(total, pageSize),
+      },
+    };
+  }
+
+  /**
+   * Add or update a user's role in a given shop.
+   * Accepts shop_admin or shop_staff (default: shop_admin).
+   * Idempotent via `upsert` on composite key `(user_id, shop_id)`.
+   */
+  async addShopAdmin(shopId: string, dto: AddShopAdminDto, actorUserId?: string | null) {
+    await this.crud.findOne(shopId); // throws 404 if shop does not exist
+
+    if (dto.user_id && !isUUID(dto.user_id)) {
+      throw new AppException(ErrorCode.VALIDATION_ERROR, 'user_id must be a valid UUID.');
+    }
+
+    const assignedRole: ShopRole = dto.role ?? ShopRole.shop_admin;
+
+    const user =
+      dto.user_id != null
+        ? await this.prisma.user.findUnique({ where: { id: dto.user_id } })
+        : dto.email != null
+          ? await this.prisma.user.findUnique({ where: { email: dto.email } })
+          : null;
+
+    if (!user) {
+      if (!dto.email || !dto.password) {
+        throw new AppException(
+          ErrorCode.NOT_FOUND,
+          'User not found. Provide `password` to create the account when user does not exist.',
+        );
+      }
+
+      const password_hash = await bcrypt.hash(dto.password, 10);
+      return this.prisma.$transaction(async (tx) => {
+        const created = await tx.user.create({
+          data: {
+            email: dto.email!,
+            password_hash,
+            full_name: dto.full_name,
+          },
+        });
+
+        const upserted = await tx.userShopRole.upsert({
+          where: { user_id_shop_id: { user_id: created.id, shop_id: shopId } },
+          create: { user_id: created.id, shop_id: shopId, role: assignedRole },
+          update: { role: assignedRole },
+        });
+
+        const safeMetadata = JSON.stringify({ email: dto.email, created_user: true, role: assignedRole });
+        await tx.shopAuditLog.create({
+          data: {
+            shop_id: shopId,
+            actor_user_id: actorUserId ?? null,
+            action: ShopAuditAction.set_shop_member_role,
+            target_type: 'user',
+            target_id: created.id,
+            metadata_json: safeMetadata,
+          },
+        });
+
+        // Invalidate the shared shop-roles cache so the new user's role is enforced
+        // immediately on the next request rather than after the 30-second TTL expires.
+        RolesGuard.invalidateShopRolesCache(created.id);
+        return upserted;
+      });
+    }
+
+    const upserted = await this.prisma.userShopRole.upsert({
+      where: { user_id_shop_id: { user_id: user.id, shop_id: shopId } },
+      create: { user_id: user.id, shop_id: shopId, role: assignedRole },
+      update: { role: assignedRole },
+    });
+    await this.audit.createAuditLog(
+      shopId,
+      ShopAuditAction.set_shop_member_role,
+      actorUserId ?? null,
+      'user',
+      user.id,
+      { email: dto.email ?? null, created_user: false, role: assignedRole },
+    );
+    // Invalidate the shared shop-roles cache so the updated role is enforced immediately.
+    RolesGuard.invalidateShopRolesCache(user.id);
+    return upserted;
+  }
+
+  /**
+   * Set password for a user who belongs to the shop (super-admin only, via controller).
+   */
+  async resetMemberPassword(
+    shopId: string,
+    memberUserId: string,
+    plainPassword: string,
+    actorUserId?: string | null,
+  ) {
+    await this.crud.findOne(shopId);
+    const membership = await this.prisma.userShopRole.findUnique({
+      where: {
+        user_id_shop_id: { user_id: memberUserId, shop_id: shopId },
+      },
+    });
+    if (!membership) {
+      throw new AppException(
+        ErrorCode.NOT_FOUND,
+        'User is not a member of this shop.',
+      );
+    }
+    const password_hash = await bcrypt.hash(plainPassword, 10);
+    await this.prisma.user.update({
+      where: { id: memberUserId },
+      data: { password_hash },
+    });
+    await this.audit.createAuditLog(
+      shopId,
+      ShopAuditAction.reset_member_password,
+      actorUserId ?? null,
+      'user',
+      memberUserId,
+      { reset: true },
+    );
+    return { updated: true };
+  }
+
+  /**
+   * Enable or disable login for a user who belongs to the shop.
+   */
+  async setMemberAccountActive(
+    shopId: string,
+    memberUserId: string,
+    is_active: boolean,
+    actorUserId?: string | null,
+  ) {
+    await this.crud.findOne(shopId);
+    const membership = await this.prisma.userShopRole.findUnique({
+      where: {
+        user_id_shop_id: { user_id: memberUserId, shop_id: shopId },
+      },
+    });
+    if (!membership) {
+      throw new AppException(
+        ErrorCode.NOT_FOUND,
+        'User is not a member of this shop.',
+      );
+    }
+    await this.prisma.user.update({
+      where: { id: memberUserId },
+      data: { is_active },
+    });
+    await this.audit.createAuditLog(
+      shopId,
+      ShopAuditAction.set_member_active,
+      actorUserId ?? null,
+      'user',
+      memberUserId,
+      { is_active },
+    );
+    return { updated: true };
+  }
+}
